refactor: create redis lib to fix circular dependency
