+++ libs/redis/index.ts
@@ -1,0 +1,3 @@
+export { RedisModule } from './redis.module';
+export { RedisService } from './redis.service';
+
